directory_source_ifs test to group 3
